pt sorting now handled by backend and populates on dash
