+++ app/(tabs)/incident_command.tsx
@@ -7,30 +7,16 @@ import ParallaxScrollView from '@/components/ParallaxScrollView';
 import { IconSymbol } from '@/components/ui/IconSymbol';
 import { Patient, ColorScheme } from '@/components/lib/types';
 import  PatientTable  from '@/components/ic_dash/pt_table';
+import  PriorityCards  from '@/components/ic_dash/priority_cards';
 import  StaffList  from '@/components/ic_dash/staff_list';
+import { useQueryPatients } from '@/components/hooks/use-query-patients';
 
-import  { triageStatusColors, initialPatients, zoneStaff }  from '@/components/ic_dash/constants'
+
+import  {zoneStaff }  from '@/components/ic_dash/constants'
 
 
 
 export default function IncidentCommandDashboard() {
-  const [patients, setPatients] =  useState<Patient[]>(initialPatients);
-
-  // Calculate patient counts by priority
-  const priorityCounts = patients.reduce<Record<string, number>>((counts, patient) => {
-    const priority = patient.triageStatus;
-    counts[priority] = (counts[priority] || 0) + 1;
-    return counts;
-  }, {});
-
-
-  // Render priority status card
-  const renderPriorityCard = (title: string, count = 0, colorScheme: ColorScheme) => (
-    <View style={[styles.priorityCard, { backgroundColor: colorScheme.bg }]}>
-      <Text style={[styles.priorityCount, { color: colorScheme.value }]}>{count || 0}</Text>
-      <Text style={[styles.priorityLabel, { color: colorScheme.text }]}>{title}</Text>
-    </View>
-  );
 
   return (
     <ThemedView style={styles.container}>
@@ -42,12 +28,7 @@ export default function IncidentCommandDashboard() {
         {/* </ThemedView> */}
 
       <ThemedView style={styles.priorityCardsContainer}>
-        <ScrollView horizontal >
-          {renderPriorityCard('IMMEDIATE', priorityCounts['Immediate'], triageStatusColors.Immediate)}
-          {renderPriorityCard('DELAYED', priorityCounts['Delayed'], triageStatusColors.Delayed)}
-          {renderPriorityCard('MINOR', priorityCounts['Minor'], triageStatusColors.Minor)}
-          {renderPriorityCard('EXPECTANT', priorityCounts['Expectant'], triageStatusColors.Expectant)}
-        </ScrollView>
+        <PriorityCards></PriorityCards>
       </ThemedView>
       </ThemedView>
       <ThemedView style={styles.mainBody}>
@@ -105,28 +86,7 @@ const styles = StyleSheet.create({
     justifyContent: 'flex-end', // Aligns cards to the right
     alignItems: 'center', // Ensures proper vertical alignment
   },
-  priorityCard: {
-    width: 120,
-    height: 80,
-    borderRadius: 8,
-    padding: 8,
-    justifyContent: 'center',
-    alignItems: 'center',
-    marginRight: 12,
-    shadowColor: '#000',
-    shadowOffset: { width: 0, height: 1 },
-    shadowOpacity: 0.2,
-    shadowRadius: 2,
-    elevation: 2,
-  },
-  priorityCount: {
-    fontSize: 32,
-    fontWeight: 'bold',
-  },
-  priorityLabel: {
-    fontSize: 14,
-    textAlign: 'center',
-  },
+  
   mainBody: {
     flexDirection: 'row',
     alignItems: 'flex-start',
