+++ components/lib/types.ts
@@ -12,7 +12,7 @@ export type PatientStatus =
   | "Transport Complete";
 
 // Define the available triage statuses
-export type TriageStatus = "Minor" | "Delayed" | "Immediate" | "Expectant";
+export type TriageStatus = "Minor" | "Delayed" | "Immediate" | "Expectant" | null;
 
 export type Patient = {
   // _id and _createdAt are generated by Convex when a new patient is added to the DB
